adjust sst1m eventsource to novel protozfits interface (#668)
* use new interface of protozfits v0.44.3

* install new protozfits interface

* try to distinguish between sst1m and other zfits files

The novel protozfits interface SimpleFile is more general than the interface in v0.44.1.
This means that operations that were previously hidden in protozfitsreader were now shifted into the eventsource in ctapipe. I think this makes more sense, since it is much more explicit.

A reader should, after all, only read the data from the file and not apply some arbitrary transformations on the data, however common they might be...

For example, it is necessary to sort the waveforms, since they are not stored in the order of pixel_ids but in some other order. Previously this sorting was done in the protozfitsreader, now it is done in the event source.